Moved nav bar markup, code to new NavigationBar component
