+++ src/App.js
@@ -1,19 +1,17 @@
 import React, { Component } from 'react';
-import { Router, Route, Link } from 'react-router-dom';
+import { Router, Route } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { Label } from 'office-ui-fabric-react';
-import { IconButton } from 'office-ui-fabric-react/lib/Button';
 
 import { history } from './_helpers';
 import { alertActions } from './_actions';
 import { PrivateRoute } from './_components';
+import { NavigationBar } from './NavigationBar';
 import { HomePage } from './HomePage/HomePage';
 import { LoginPage } from './LoginPage/LoginPage';
 import { RegisterPage } from './RegisterPage/RegisterPage';
 import Pricing from './PricingPage/Pricing';
 import { LogFormLayout } from './LogPage/LogFormLayout';
 import { LogList } from './LogListPage/LogListPage';
-
 
 import './App.css';
 
@@ -28,44 +26,8 @@ class App extends Component {
     });
   }
 
-  async isLoggedIn() {
-    // let u = localStorage.getItem('user');    
-    // return (u !== null) ? true : false;
-
-    
-  }
-
-  componentDidMount() {
-    
-  }
-
   render() {
     const { alert, users } = this.props;
-
-    const styles = {
-      navBar: {
-        display: 'flex',
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        backgroundColor: '#484644',
-
-        left: {
-          display: 'flex',
-          flexDirection: 'row'
-        },
-        right: {
-          display: 'flex',
-          flexDirection: 'row'
-        }
-      },
-      linkStyle: {
-        textDecoration: 'none',
-        margin: '0 0.25em 0 0.25em'
-      },
-      whiteLabel: {
-        color: 'white'
-      }
-    }
 
     return (
       <div className="App">
@@ -73,43 +35,7 @@ class App extends Component {
           alert.message && <div className={`alert ${alert.type}`}>{alert.message.message}</div>
         }
         <Router history={history}>
-            <div style={styles.navBar}>
-                <div style={styles.navBar.left}>
-                  <IconButton 
-                    style={{color: 'white'}} 
-                    iconProps={{ iconName: 'CollapseMenu' }} 
-                    title="CollapseMenu" 
-                    ariaLabel="CollapseMenu" />
-                  <Link to="/fuel-log" style={styles.linkStyle}>
-                    <Label style={styles.whiteLabel}>Add Log</Label>
-                  </Link>
-                  <Link to="/list" style={styles.linkStyle}>
-                    <Label style={styles.whiteLabel}>Logs</Label>
-                  </Link>
-                  <Link to="/pricing" style={styles.linkStyle}>
-                    <Label style={styles.whiteLabel}>Pricing</Label>
-                  </Link>
-                </div>
-
-                <div style={styles.navBar.right}>
-
-                  {
-                    users && users.user ? 
-                    <Link to="/login" style={styles.linkStyle}>
-                      <Label style={styles.whiteLabel}>Logout</Label>
-                    </Link> :
-                    <div>
-                      <Link to="/login" style={styles.linkStyle}>
-                        <Label style={styles.whiteLabel}>Sign In</Label>
-                      </Link>
-                      <Link to="/register" style={styles.linkStyle}>
-                        <Label style={styles.whiteLabel}>Register</Label>
-                      </Link>
-                    </div>
-                  }
-                  
-                </div>
-            </div>
+            <NavigationBar />
 
             <div>
                 <PrivateRoute exact path="/" component={HomePage} />
