+++ src/LoginPage/LoginPage.js
@@ -1,8 +1,4 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import { connect } from 'react-redux';
-import { userActions } from '../_actions';
-
 import { LoginFormik } from './LoginFormik';
 import styles from '../styling/styles';
